# swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Coffee Supply Chain",
        "description": "Coffee Supply Chain Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/supplier/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of supplier that we want to find",
                    "type": "integer"
                }
            ],
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get all Suppliers from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "body",
                        "description": "Supplier that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                ],
                "summary": "Get all Suppliers from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                }
            }
        },
        "/supplier": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get all Suppliers from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "body",
                        "description": "Supplier that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                ],
                "summary": "Get all Suppliers from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                }
            }
        },
        "/shipment/{shipment_id}": {
            "parameters": [
                {
                    "name": "shipment_id",
                    "in": "path",
                    "required": true,
                    "description": "ID of shipment that we want to find",
                    "type": "integer"
                }
            ],
            "get": {
                "tags": [
                    "Shipment"
                ],
                "summary": "Get all Shipment from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Shipment_id"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Shipment"
                ],
                "parameters": [
                    {
                        "name": "Shipment",
                        "in": "body",
                        "description": "Shipment that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Shipment_id"
                        }
                    }
                ],
                "summary": "Get all Shipments from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Shipment_id"
                        }
                    }
                }
            }
        },
        "/shipment": {
            "get": {
                "tags": [
                    "Shipment"
                ],
                "summary": "Get all Shipment from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Shipment"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Shipment"
                ],
                "parameters": [
                    {
                        "name": "Shipment",
                        "in": "body",
                        "description": "Supplier that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Shipment"
                        }
                    }
                ],
                "summary": "Get all Shipment from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Shipment_id"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Producers": {
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "isPublic": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "books": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "amount": {
                                "type": "number"
                            }
                        }
                    }
                },
                "companies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "Producers_id": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "region_id": {
                    "type": "integer"
                }
            }
        },
...
        "Shipment": {
            "required": ["origin_id", "destination_id", "quantity", "status"],
            "properties": {
                "origin_id": {
                    "type": "integer"
                },
                "destination_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "Shipment_id": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "region_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
          "name": "Authorization",
          "in": "header",
          "type": "apiKey",
          "description": "JWT Authorization header"
        }
      },
      "security": [ { "bearerAuth": [] } ]
}